# EcommerceAPI.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "EcommerceAPI",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/category/createCategory": {
      "post": {
        "summary": "create category",
        "tags": [
          "category"
        ],
        "operationId": "createcategory",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createcategoryrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/category/getCategory": {
      "get": {
        "summary": "get category",
        "tags": [
          "category"
        ],
        "operationId": "getcategory",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/category/getCategory/1": {
      "get": {
        "summary": "get category id",
        "tags": [
          "category"
        ],
        "operationId": "getcategoryid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/category/updateCategory/1": {
      "put": {
        "summary": "update category",
        "tags": [
          "category"
        ],
        "operationId": "updatecategory",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updatecategoryrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "delete category",
        "tags": [
          "category"
        ],
        "operationId": "deletecategory",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/category/getCategoryActive": {
      "get": {
        "summary": "get category active",
        "tags": [
          "category"
        ],
        "operationId": "getcategoryactive",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/flashSale/createFlashSale": {
      "post": {
        "summary": "create flash sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "createflashsale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createflashsalerequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/getFlashSale": {
      "get": {
        "summary": "get flash Sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "getflashSale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/getFlashSale/1": {
      "get": {
        "summary": "get flash Sale ID",
        "tags": [
          "flash sale"
        ],
        "operationId": "getflashSaleID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/updateFlashSale/1": {
      "put": {
        "summary": "update flash sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "updateflashsale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateflashsalerequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/flashSale/notification": {
      "post": {
        "summary": "notification",
        "tags": [
          "flash sale"
        ],
        "operationId": "notification",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/image/createImage": {
      "post": {
        "summary": "create image",
        "tags": [
          "image"
        ],
        "operationId": "createimage",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "img",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "itemId",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/image/updateImage/1": {
      "put": {
        "summary": "update image",
        "tags": [
          "image"
        ],
        "operationId": "updateimage",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "img",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "itemId",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createcategoryrequest": {
      "title": "createcategoryrequest",
      "example": {
        "categoryName": "do an",
        "location": 3,
        "status": 1
      },
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string"
        },
        "location": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "categoryName",
        "location",
        "status"
      ]
    },
    "updatecategoryrequest": {
      "title": "updatecategoryrequest",
      "example": {
        "categoryName": "công nghệ",
        "location": 5,
        "status": 1
      },
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string"
        },
        "location": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "categoryName",
        "location",
        "status"
      ]
    },
    "createflashsalerequest": {
      "title": "createflashsalerequest",
      "example": {
        "flashSaleName": "FS8-3",
        "startTime": "2021-07-15T03:30:00",
        "stopTime": "2021-07-15T04:55:00",
        "discount": 10,
        "amount": 12,
        "idItems": [
          1,
          2,
          3
        ]
      },
      "type": "object",
      "properties": {
        "flashSaleName": {
          "type": "string"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "idItems": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "required": [
        "flashSaleName",
        "startTime",
        "stopTime",
        "discount",
        "amount",
        "idItems"
      ]
    },
    "updateflashsalerequest": {
      "title": "updateflashsalerequest",
      "example": {
        "flashSaleName": "Fs20-11",
        "startTime": "2021-07-15T03:30:00",
        "stopTime": "2021-07-15T04:55:00",
        "discount": 10,
        "amount": 12,
        "idItems": [
          1,
          2,
          3
        ]
      },
      "type": "object",
      "properties": {
        "flashSaleName": {
          "type": "string"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "idItems": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "required": [
        "flashSaleName",
        "startTime",
        "stopTime",
        "discount",
        "amount",
        "idItems"
      ]
    },
...
  },
...
}